+++ api-testing/multicloud-api/src/main/java/com/io/sodafoundation/utils/Constant.java
@@ -13,19 +13,11 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.soda.utils;
+package com.io.sodafoundation.utils;
 
 /**
- * This class used for logs print.
+ * In this class set Inputs path.
  */
-public class Logger {
-    /**
-     * Print log
-     *
-     * @param message log message
-     * @param <T>     data type
-     */
-    public static <T>  void log(T message) {
-        System.out.println(message);
-    }
+public class Constant {
+    public static final  String PATH = System.getenv("INPUT_PATH")+"/src/main/resources/inputs/";
 }
