+++ api-testing/multicloud-api/src/main/java/com/io/sodafoundation/utils/ConstantUrl.java
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.soda.utils;
+package com.io.sodafoundation.utils;
 
 /**
  *  In this class set URL
