+++ api-testing/multicloud-api/src/main/java/com/io/sodafoundation/utils/Utils.java
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.soda.utils;
+package com.io.sodafoundation.utils;
 
 import java.io.File;
 import java.io.IOException;
